third_party/WebKit/LayoutTests/media/video-controls-captions-on-off-expected.txt

Tests that tracks can be turned on and off through the track selection menu

EVENT(canplaythrough)

** Caption button should be visible and enabled.
EXPECTED (captionsButtonCoordinates[0] > '0') OK
EXPECTED (captionsButtonCoordinates[1] > '0') OK
EXPECTED (captionsButtonElement.disabled == 'false') OK

** The captions track should be listed in textTracks, but not yet loaded. **
EXPECTED (video.textTracks.length == '1') OK
EXPECTED (video.textTracks[0].mode == 'hidden') OK
No text track cue with display id '-webkit-media-text-track-display' is currently visible
No text track cue with display id '-webkit-media-text-track-display' is currently visible
No text track cue with display id '-webkit-media-text-track-display' is currently visible

** Captions track should become visible after the track is selected **
*** Click the CC button.
EXPECTED (displayElement.innerText == 'First') OK
EXPECTED (displayElement.innerText == 'Second') OK
EXPECTED (displayElement.innerText == 'Third') OK

** Captions should not be visible after they're turned off through the menu **
*** Click the CC button.
No text track cue with display id '-webkit-media-text-track-display' is currently visible
No text track cue with display id '-webkit-media-text-track-display' is currently visible
No text track cue with display id '-webkit-media-text-track-display' is currently visible

** Captions track should become visible after the track is selected again **
*** Click the CC button.
EXPECTED (displayElement.innerText == 'First') OK
EXPECTED (displayElement.innerText == 'Second') OK
EXPECTED (displayElement.innerText == 'Third') OK

END OF TEST
